% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{citations}
\alias{citations}
\title{Returns citations for the datasets in MADTRAITS}
\usage{
citations(x, format = c("bibtex", "ris", "citeproc", "schema.org"))
}
\arguments{
\item{x}{The MADtraits database object for which you want citations}

\item{format}{One of "bibtex", "ris", "citeproc", or "schema.org",
which is the format in which your citations will be
returned. Default is "bibtex"}
}
\value{
Character vector of citations in chosen format.
}
\description{
This will generate citations (in BibTeX format) for a given MADTRAITS
dataset.
}
\author{
Will Pearse; Sylia Kinosian
}
