% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean.MADtraits}
\alias{clean.MADtraits}
\title{Cleaning MADtraits objects}
\usage{
clean.MADtraits(x)
}
\arguments{
\item{x}{\code{\link{MADtraits}} object}
}
\value{
\code{\link{MADtraits}} object
}
\description{
A quick method that looks for obvious issues (capitalisation, known
mis-spellings, etc.) within the raw data that MADtraits downloads. This
is a *strongly* recommend secibd step for anyone working with the
output from MADtraits.
}
\note{
See \code{\link{convertunits.MADtraits}} and
    \code{\link{taxonlookup.MADtraits}} for functions to match
    variable units and taxonomic names in more sophisticated ways
}
\examples{
# Grab some example data (you should work with the output from the MADtraits function)
demo <- .cavenderbares.2015a()
demo <- clean.MADtraits(demo)
}
\seealso{
convertunits.MADtraits taxonlookup.MADtraits
}
\author{
Will Pearse
}
