% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean.natdb}
\alias{clean.natdb}
\title{Cleaning NATDB objects}
\usage{
clean.natdb(x)
}
\arguments{
\item{x}{\code{\link{natdb}} object}
}
\value{
\code{\link{natdb}} object
}
\description{
A quick method that looks for obvious issues (capitalisation, known
mis-spellings, etc.) within the raw data that NATDB downloads. This
is a *strongly* recommend secibd step for anyone working with the
output from NATDB.
}
\note{
See \code{\link{convert.natdb.units}} and
    \code{\link{lookup.natdb.names}} for functions to match
    variable units and taxonomic names in more sophisticated ways
}
\examples{
# Grab some example data (you should work with the output from the natdb function)
demo <- .cavenderbares.2015a()
demo <- convert.natdb.units(demo)
}
\seealso{
convert.natdb.units lookup.natdb.names
}
\author{
Will Pearse
}
