% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{lookup.natdb.names}
\alias{lookup.natdb.names}
\title{Harmonising taxonomic names within a NATDB object}
\usage{
lookup.natdb.names(x, thresh = 0.8, use.cache = TRUE)
}
\arguments{
\item{x}{\code{\link{natdb}} object}

\item{thresh}{Threshold of certainty to be used as a minimum when
assigning new names to a species. The default, of 0.8, has not
been chosen with any particular intelligence.}

\item{use.cache}{whether to use the internal cache of species names
that NATDB ships with. This is generated by Will Pearse on an
ad hoc basis; it's probably right, but please don't sue me if
it isn't! :D}
}
\value{
\code{\link{natdb}} object
}
\description{
A *very* light wrapper around \code{\link[taxize]{gnr_resolve}} to
harmonise taxonomic names across datasets. Taxonomy is a difficult
thing to get right, and many would argue it is philosophically
impossible to get right unless you verify the taxonomy of specimens
with vouchers. A bit of careful checking of the original values
*before* doing taxonomic cleaning is often beneficial.
}
\note{
*Please* run your data through \code{\link{clean.natdb}}
    before running this, as it will neaten up obvious mistakes in
    the input data's species names.

*Please* consider using the built-in taxonomic lookup table, or
subsetting your data to the groups/traits of interest before
running the lookup. It can take a *tremendously* long time to
lookup and resolve all ~110,000 species for which NATDB downloads
data, and you're not going to enjoy the process of waiting!
}
\examples{
# Grab some example data (you should work with the output from the natdb function)
demo <- .pearse.2014()
# Convert all the units to the most common type
# - (a bit silly here as they're all standardised anyway!)
demo <- convert.natdb.units(demo)
# Now convert the "diameter" from millimeters to meters
units <- setNames("m", "diameter")
# (this is the same as the following)
units <- "m"
names(units) <- "diameter"
demo <- convert.natdb.units(demo, units)
#...you don't have to clean up just one unit
# - your "units" vector can be as long as you want
# Grab some example data (you should work with the output from the natdb function)
demo <- .cavender_bares.2015a()
# - before using it, clean up the data to remove obvious differences
demo <- clean.natdb(demo)
# Now run a thorough taxonomic check
demo <- lookup.natdb.names(demo, use.cache=FALSE)
#...in this case with only six species it was fast
}
\seealso{
clean.natdb convert.natdb.units
}
\author{
Will Pearse and Kathryn M. Welglarz

Will Pearse and Mallory Hagadorn
}
