% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADtraits.R
\name{MADtraits}
\alias{MADtraits}
\alias{print.MADtraits}
\alias{summary.MADtraits}
\alias{[.MADtraits}
\alias{species}
\alias{traits}
\alias{as.data.frame.MADtraits}
\title{Make A Database of Traits}
\usage{
MADtraits(cache, datasets, delay = 5)

\method{print}{MADtraits}(x, ...)

\method{summary}{MADtraits}(object, ...)

\method{[}{MADtraits}(x, spp, traits)

species(x, ...)

traits(x, ...)

\method{as.data.frame}{MADtraits}(
  x,
  row.names,
  optional,
  num.func = mean,
  cat.func = function(x) names(which.max(table(x))),
  data = 10,
  ...
)
}
\arguments{
\item{cache}{Specify an existing directory/folder where datasets
can be downloaded to and stored. If a dataset is already
present in this directory, it will not be downloaded from the
server but instead loaded locally. We *STRONGLY* advise you to
specify a cache location.}

\item{datasets}{Character vector of datasets to be searched for
trait data. If not specified (the default) all trait datasets
will be downloaded and returned.}

\item{delay}{How many seconds to wait between downloading and
processing each dataset (default: 5). This delay may seem
large, but if you specify a \code{cache} (see above) you only
need do it once, and specifying a large delay ensures you don't
over-stretch servers. Keeping servers happy is good for you
(they won't reject you!) and good for them (they can help
everyone).}

\item{x}{\code{MADtraits} object to be printed}

\item{...}{ignored}

\item{object}{\code{MADtraits} object to be summarised}

\item{spp}{\code{character} vector of species to subset the
\code{MADtraits} object down to}

\item{traits}{\code{character} vector of traits to subset the
\code{MADtraits} object down to}

\item{row.names}{Ignored}

\item{optional}{Ignored}

\item{num.func}{To summarise data at the species level (which is
done by default), a function is needed to summarise the
continuous data at the species level. This argument specifies
this function; while the default is \code{\link[base]{mean}},
you could change this to \code{\link[stats]{median}} or specify
your own standard error function.}

\item{cat.func}{To summarise data at the species level (which is
done by default), a function is needed to summarise the
continuous data at the species level. This argument specifies
this function; while the default is to just return the modal
value, other options could be used.}

\item{data}{Which variables to summarise into a data.frame. If
numeric, the top \code{data} variables with the most data will
be summarised (default: top ten variables). You can also
specify variable names if you wish. Note: there are over a
thousand variables and over 100,000 species in MADTRAITS - you *do
not* want to turn all of this data into a single
\code{data.frame}, as it will be unmanageable. Subset your data
down first to your species or traits of interest (see
"examples")}
}
\value{
MADtraits.data object.
}
\description{
The key function of the MADtraits package. When run with defaults,
it will download and build a database of species' traits from all
the manuscript sources in the package. *Please* make
use the the \code{cache} feature, as it will massively speed and
ease your use of the package.
}
\examples{
\dontrun{
# Download all MADtraits data, and cache (save) it on your hard-drive for use later
data <- MADtraits(cache="Documents/MADtraits/cache")
# Perform basic checks and cleaning on that data
clean.data <- clean.MADtraits(data)

# Subset data down to the traits you want (notice the comma position)
subset.data <- clean.data[,c("specific_leaf_area","height")]
# Subset data down to the species you want (notice the comma position)
subset.data <- clean.data[c("quercus_robur","quercus_ilex"),]
# Subset multiple things at once
clean.data[c("quercus_robur","quercus_ilex"),"specific_leaf_area"]

# Convert that into a data.frame for use in an analysis
neat.data <- as.data.frame(clean.data)
}
}
\seealso{
clean.MADtraits convert.MADtraits.units lookup.MADtraits.species
}
\author{
Will Pearse
}
