% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MADtraits_datasets}
\alias{MADtraits_datasets}
\title{Get vector of the IDs for all datasets available within
    MADtraits}
\usage{
MADtraits_datasets()
}
\value{
Vector of dataset IDs
}
\description{
Returns a vector of all IDs for the datasets that
    MADtraits can download. This can be used to identify particular
    datasets you wish to download using \code{\link{MADtraits}}.
}
\examples{
\dontrun{
datasets <- MADtraits_datasets()
data <- MADtraits(datasets = datasets[1:2])
}

}
\seealso{
citations MADtraits
}
\author{
Will Pearse; Matthew Helmus
}
\keyword{datasets}
