% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natdb.R
\name{natdb}
\alias{natdb}
\title{Builds a trait database}
\usage{
natdb(cache, datasets, delay = 5)
}
\arguments{
\item{cache}{Specify an existing directory/folder where datasets
can be downloaded to and stored. If a dataset is already
present in this directory, it will not be downloaded from the
server but instead loaded locally. We *STRONGLY* advise you to
specify a cache location.}

\item{datasets}{Character vector of datasets to be searched for
trait data. If not specified (the default) all trait datasets
will be downloaded and returned.}

\item{delay}{How many seconds to wait between downloading and
processing each dataset (default: 5). This delay may seem
large, but if you specify a \code{cache} (see above) you only
need do it once, and specifying a large delay ensures you don't
over-stretch servers. Keeping servers happy is good for you
(they won't reject you!) and good for them (they can help
everyone).}
}
\value{
natdb.data object. XXX
}
\description{
The key function of the natdb package. When run with defaults, it
will download and build a database of species' traits from all the
manuscript sources in the package. This totals XXX
manuscripts/databases, XXX species, and XXX traits. *Please* make
use the the \code{cache} feature, as it will massively speed and
ease your use of the package.
}
\author{
Will Pearse; USU Biology Nerd Group (XXX)
#@examples
# Limit the scope of these as they have to work online on servers!...
# Also give an example (notrun) of how to use the cache
#@seealso
}
