% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\docType{data}
\name{natdb_citations}
\alias{natdb_citations}
\title{Citations for all datasets within NATDB}
\format{\code{natdb_citations} contains a \code{data.frame} of
    citations and paper information.}
\usage{
natdb_citations
}
\description{
Citations for all datasets within NATDB. These *MUST* be
    cited when publishing anything that makes use of this package:
    cite the data in preference to this package, if needs be. You
    will rarely use this data directly; instead, use the
    \code{\link{citations}} function.
}
\seealso{
citations
}
\keyword{datasets}
