% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{natdb_datasets}
\alias{natdb_datasets}
\title{Get vector of the IDs for all datasets available within NATDB}
\usage{
natdb_datasets()
}
\value{
Vector of dataset IDs
}
\description{
Returns a vector of all IDs for the datasets that NATDB can download. This can be used to identify particular datasets you wish to download using \code{\link{natdb}}.
}
\examples{
\dontrun{
datasets <- natdb_datasets()
data <- natdb(datasets = datasets[1:2])
}

}
\seealso{
citations natdb
}
\author{
Will Pearse; Matthew Helmus
}
\keyword{datasets}
