% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citations.R
\name{citations}
\alias{citations}
\title{Returns citations for the datasets in NATDB}
\usage{
citations(x)
}
\arguments{
\item{x}{The NATDB database object for which you want citations}
}
\value{
Character vector of citations in BibTeX format
}
\description{
This will generate citations (in BibTeX format) for a given NATDB
dataset.
}
\author{
Will Pearse; Sylia Kinosian
#@examples
# Limit the scope of these as they have to work online on servers!...
# Also give an example (notrun) of how to use the cache
#@seealso
}
