% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natdb.R
\name{natdb}
\alias{natdb}
\title{Builds a trait database}
\usage{
natdb(datasets, cache, delay = 5)
}
\arguments{
\item{datasets}{Character vector of datasets to be searched for
trait data. If not specified (the default) all trait datasets
will be downloaded and returned.}

\item{species}{Character vector of species to be searched for trait
data. If not specified (the default) data for all species will
be downloaded and returned.}

\item{traits}{Character vector of traits to be searched for
data. If not specified (the default) data for all traits will
be downloaded and returned.}
}
\value{
natdb.data object. XXX
}
\description{
The key function of the natdb package. When run with defaults, it
will download and build a database of species' traits from all the
manuscript sources in the package. This totals XXX
manuscripts/databases, XXX species, and XXX traits. Please note
that all parameters are interactive; thus specifying \code{species}
and \code{traits} constraints will constraint according to both,
for example. Please also note that specifying any kind of
constraints makes use of the package's built-in cache of what
species and traits information are available in each database;
making use of this on the GitHub (developer) build of this package
is not advisable, and (further) it is impossible for us to verify
whether the datasets NATDB searches have been updated since the
package was last built.
}
\author{
Will Pearse; USU Biology Nerd Group (XXX)
#@examples
# Limit the scope of these as they have to work online on servers!...
#@seealso
}
